Correctly rotate triangle faces for cube
We have working transformations of triangle faces,
but we are still missing some important things. Namely,
there is no depth buffer and so we just render whatever
triangle happens to be last in our collection on top of
previously rendered tris. Bad! Also, I think my
coordinates for the trianges are not quite right - they
are not forming square faces.